chore: cleanup unneeded contexts
