--- src/contexts/Catalog/index.ts
@@ -1,1 +1,0 @@
-export { useCatalog, CatalogProvider } from "./Catalog";
