--- src/contexts/Stats/index.ts
@@ -1,1 +1,0 @@
-export { useStats, StatsProvider } from "./Stats";
