--- src/contexts/Versions/index.ts
@@ -1,1 +1,0 @@
-export { useVersions, VersionsProvider } from "./Versions";
