+++ src/hooks/useCatalog/index.ts
@@ -1,0 +1,1 @@
+export * from "./useCatalog";
