+++ src/index.tsx
@@ -5,14 +5,11 @@ import { BrowserRouter as Router } from "react-router-dom";
 import "./index.css";
 import { App } from "./App";
 import { AnalyticsProvider } from "./contexts/Analytics";
-import { CatalogProvider } from "./contexts/Catalog";
 import { ConfigProvider } from "./contexts/Config";
 import { ExternalLinkWarningProvider } from "./contexts/ExternalLinkWarning";
 import { SearchProvider } from "./contexts/Search";
 import { ShortbreadProvider } from "./contexts/Shortbread";
-import { StatsProvider } from "./contexts/Stats";
 import { Theme } from "./contexts/Theme";
-import { VersionsProvider } from "./contexts/Versions";
 import { unregister } from "./register-service-worker";
 import { reportWebVitals } from "./reportWebVitals";
 
@@ -25,19 +22,13 @@ ReactDOM.render(
         <ShortbreadProvider>
           <AnalyticsProvider>
             <ConfigProvider>
-              <CatalogProvider>
-                <StatsProvider>
-                  <VersionsProvider>
-                    <SearchProvider>
-                      <Theme>
-                        <ExternalLinkWarningProvider>
-                          <App />
-                        </ExternalLinkWarningProvider>
-                      </Theme>
-                    </SearchProvider>
-                  </VersionsProvider>
-                </StatsProvider>
-              </CatalogProvider>
+              <SearchProvider>
+                <Theme>
+                  <ExternalLinkWarningProvider>
+                    <App />
+                  </ExternalLinkWarningProvider>
+                </Theme>
+              </SearchProvider>
             </ConfigProvider>
           </AnalyticsProvider>
         </ShortbreadProvider>
