+++ src/views/Package/PackageHeader/Details.tsx
@@ -20,8 +20,8 @@ import { LicenseLink } from "../../../components/LicenseLink";
 import { NavLink } from "../../../components/NavLink";
 import { Time } from "../../../components/Time";
 import { FORMATS } from "../../../constants/dates";
-import { useStats } from "../../../contexts/Stats";
 import { useConfigValue } from "../../../hooks/useConfigValue";
+import { useStats } from "../../../hooks/useStats";
 import { highlightsFrom } from "../../../util/package";
 import { getRepoUrlAndHost, getSearchPath } from "../../../util/url";
 import { usePackageState } from "../PackageState";
